+++ src/compas_rcf/utils/docker/docker_cmds.py
@@ -4,6 +4,10 @@ from __future__ import print_function
 
 import os
 from subprocess import run
+import logging
+
+log = logging.getLogger(__name__)
+
 
 RUN_KWARGS = {
     "check": True,
@@ -16,7 +20,7 @@ RUN_KWARGS = {
 }
 
 
-def setup_env_vars(env_vars):
+def _setup_env_vars(env_vars):
     list_vars = []
     for key in env_vars:
         if os.name == "nt":
@@ -32,11 +36,13 @@ def compose_up(
 ):
     cmd = ["docker-compose", "--file", str(path), "up", "--detach"]
 
+    log.debug("Env vars: {}".format(env_vars))
+
     if ignore_orphans:
         env_vars.update({"COMPOSE_IGNORE_ORPHANS": "true"})
 
     if len(env_vars) > 0:
-        cmd = setup_env_vars(env_vars) + cmd
+        cmd = _setup_env_vars(env_vars) + cmd
         shell = True
     else:
         shell = False
@@ -47,40 +53,6 @@ def compose_up(
     if remove_orphans:
         cmd.append("--remove-orphans")
 
-    run(cmd, shell=shell, **RUN_KWARGS)
-
-
-def compose_run(path, **env_vars):
-    cmd = ["docker-compose", "--file", str(path), "run", "--detach"]
-
-    if len(env_vars) > 0:
-        cmd = setup_env_vars(env_vars) + cmd
-        shell = True
-    else:
-        shell = False
+    log.debug("Command to run: {}".format(cmd))
 
     run(cmd, shell=shell, **RUN_KWARGS)
-
-
-def compose_down(path):
-    cmd = ["docker-compose", "--file", str(path), "down"]
-    run(cmd, **RUN_KWARGS)
-    cmd_post = ["docker-compose", "--file", str(path), "rm"]
-    run(cmd_post, **RUN_KWARGS)
-
-
-def compose_restart(path, **env_vars):
-
-    cmd = ["docker-compose", "--file", str(path), "restart"]
-
-    if len(env_vars) > 0:
-        cmd = setup_env_vars(env_vars) + cmd
-        shell = True
-    else:
-        shell = False
-
-    run(cmd, shell=shell, **RUN_KWARGS)
-
-
-def container_restart(container_name):
-    return run("docker restart {}".format(container_name), **RUN_KWARGS)